feat: refactor, add a real case of validating db
